Simplified fragmentization
Instead of using the complex function I was using to check split text
into fragments without having preformat blocks spanning across the
fragments, the render function will instead detect them and then join
the next several fragments until it finds the closing marks.  It will
only check for such if a line starts with {{{ and the fragment with the
closing mark ends with }}}.  Inline preformat blocks must be contained
within a single fragment.

This yields a fairly decent increase in rendering speed.